build dashboard
